backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
scapy
python-nmap
pysnmp-lextudio
zeroconf
mac-vendor-lookup
